integration-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pydantic==2.5.3
pymongo==4.6.1
pdf2image==1.17.0
pytesseract==0.3.10
boto3==1.34.30
requests==2.31.0
python-dotenv==1.0.1
pytest==7.4.4
httpx==0.26.0 